+++ Gardener-main/config.h
@@ -24,8 +24,8 @@
 // PINS
 #define PIN_SENS_TEMP_HUM_INTERIOR 14
 #define PIN_SENS_TEMP_HUM_EXTERIOR 25
-#define PIN_ACT_WINDOW_0 33
-#define PIN_ACT_WINDOW_1 32
+#define PIN_ACT_WINDOW_0 12
+#define PIN_ACT_WINDOW_1 13
 #define PIN_SENS_WINDOW_ENDSTOP 39
 
 #define PIN_SENS_MOISTURE_ADDR_0 23
@@ -33,8 +33,8 @@
 #define PIN_SENS_MOISTURE_ADDR_2 18
 #define PIN_SENS_MOISTURE_ADDR_3 5
 #define PIN_SENS_MOISTURE_ENABLE 4
-#define PIN_SENS_MOISTURE_P0 12
-#define PIN_SENS_MOISTURE_P1 13
+#define PIN_SENS_MOISTURE_P0 33
+#define PIN_SENS_MOISTURE_P1 32
 
 #define PORT_FEEDER Serial2
 #define PIN_FEEDER_TX 17
@@ -49,15 +49,20 @@
 #define FEEDER_RANGE 65123
 
 // MOISTURE SENSORS
-#define SENSORS_N 12
-#define SENSORS_RESPONSE_TIMEOUT 5000
-#define SENSORS_MEASUREMENT_DURATION 5000
-#define READINGS 10
-#define FIXED_RESISTOR 4630
-#define ADC_RESOLUTION 12
-
-
-
+#define MS_MAX_SENSORS 12
+#define MS_RESPONSE_TIMEOUT 5000
+#define MS_MEASUREMENT_DURATION 5000
+#define MS_UPDATE_INTERVAL  10 * 60 * 1000
+#define MS_READINGS         10ul                // Number of measurements to take to average out noise.
+#define MS_FIXED_RESISTOR   4600l               // Resistor divider fixed resistance (Ohm).
+#define MS_ADC_RESOLUTION   12
+#define MS_VCC_VOLT         3300l               // Sensor supply voltage (mV).
+#define MS_RANGE_UPPER_VOLT 3100l               // ADC upper voltage limit (mV) for ADC_ATTEN_DB_11.
+#define MS_RANGE_LOWER_VOLT 150l                // (mV)
+#define MS_RANGE_VOLT       (MS_RANGE_UPPER_VOLT - MS_RANGE_LOWER_VOLT)
+#define MS_RANGE_UPPER_ADC  (((MS_VCC_VOLT - MS_RANGE_LOWER_VOLT) * (1 << MS_ADC_RESOLUTION)) / MS_RANGE_VOLT) // Supply voltage (ADC units)
+#define MS_RANGE_LOWER_ADC  (((0 - MS_RANGE_LOWER_VOLT) * (1 << MS_ADC_RESOLUTION)) / MS_RANGE_VOLT)             // Ground voltage in ADC units.
+#define MS_RANGE_ADC        (MS_RANGE_UPPER_ADC - MS_RANGE_LOWER_ADC)